move pickles rule wrapper to proof_systems
